@@ -4,14 +4,12 @@ version: 1.0.0
 homepage: https://github.com/bogkonstantin/flutter_ssh
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.10.0 <2.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-
-  meta: ^1.1.8
   uuid: ^3.0.0
 
 flutter:
